requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1

# Task queue
celery[redis]==5.3.6
redis==5.0.1

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.4
python-docx==1.1.0

# LlamaIndex for RAG
llama-index==0.10.12
llama-index-core==0.10.12
llama-index-embeddings-huggingface==0.1.4

# Local embeddings
sentence-transformers==2.3.1
torch==2.2.0
transformers==4.37.2
numpy<2.0.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1
